requirement.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==2.7.1
jsonpath==0.82
jsonschema==3.2.0
logger==1.4
loguru==0.4.1
lxml==4.4.1
pandas==1.3.4
pytest==7.1.1
pytest-bdd==4.1.0
pytest-html==3.1.1
PyYAML==5.4.1
request==2019.4.13
requests==2.27.1
requests-cache==0.9.5
requests-toolbelt==0.9.1
selenium==4.1.0
